finish cart

+++ src/Entity/Product.php
@@ -72,7 +72,7 @@ class Product
     private $picture_extension;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $price;
 
@@ -185,7 +185,7 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
